Widen the shifts MSVC reports as C4334
An external reporter building this branch with MSVC hit C4334, "result
of 32-bit shift implicitly converted to 64 bits", at tests/test.c:194.
The left operand was 1U, so the shift ran in 32 bits and the result was
then widened to fill a size_t. The value was never wrong, since the
count is capped at 19, but the construct is one edit away from losing
the high bits.

Casting the left operand rather than writing 1ULL is what the rest of
the tree already does, and it is the better form here: 1ULL forces
64-bit arithmetic on 32-bit targets, where size_t is 32 bits, trading
this diagnostic for a narrowing one.

Two more sites go with it. THRASH_SIZE in tests/wcet.c had the same
shape and converts to size_t at all three of its uses, so it would have
broken the build the moment the diagnostic became an error, and the
reporter would not have seen it because it never built that executable.
The static assert in src/tlsf.c divides a size_t by SL_COUNT, an
unsigned int holding a shift result, which widens the same way.

The five MSVC compile steps now pass /we4334, matching what "Avoid
unary minus on unsigned operands" did for C4146. An audit of every
shift in the MSVC-compiled translation units found no other site: the
remaining ones either already cast the left operand to size_t, or
operate on uint32_t bitmaps and stay 32-bit throughout.

+++ src/tlsf.c
@@ -180,7 +180,11 @@ TLSF_STATIC_ASSERT(sizeof(size_t) == sizeof(void *),
                    "size_t must equal pointer size");
 TLSF_STATIC_ASSERT(offsetof(tlsf_block_t, header) == BLOCK_HEADER_OFFSET,
                    "unexpected block header offset");
-TLSF_STATIC_ASSERT(ALIGN_SIZE == BLOCK_SIZE_SMALL / SL_COUNT,
+/* SL_COUNT is cast because it is an unsigned int and BLOCK_SIZE_SMALL is a
+ * size_t: without it the division widens a 32-bit shift result to 64 bits,
+ * which is what MSVC reports as C4334, and CI promotes that to an error.
+ */
+TLSF_STATIC_ASSERT(ALIGN_SIZE == BLOCK_SIZE_SMALL / (size_t) SL_COUNT,
                    "sizes are not properly set");
 TLSF_STATIC_ASSERT(BLOCK_SIZE_MIN < BLOCK_SIZE_SMALL,
                    "min allocation size is wrong");
